--- packages/components/src/stories/FocusRing/focusring.constants.ts
@@ -1,6 +1,0 @@
-/* eslint-disable implicit-arrow-linebreak */
-const DEFAULTS = {
-  SHAPE: 'mds-div' as const,
-};
-
-export { DEFAULTS };
