+++ packages/components/src/stories/FocusRing/subcomponent-focusring.constants.ts
@@ -1,0 +1,10 @@
+/* eslint-disable implicit-arrow-linebreak */
+import utils from '../../utils/tag-name';
+
+const TAG_NAME = utils.constructTagName('subcomponent-focusring');
+
+const DEFAULTS = {
+  SHAPE: 'mds-div' as const,
+};
+
+export { DEFAULTS, TAG_NAME };
