+++ packages/components/src/stories/FocusRing/subcomponent-focusring.stories.ts
@@ -1,7 +1,7 @@
 import type { Meta, StoryObj, Args } from '@storybook/web-components';
 import { html } from 'lit';
 import '.';
-import './focusring.component';
+import './subcomponent-focusring.component';
 
 const SHAPES = ['mds-div', 'mds-anchor', 'mds-button', 'mds-radio'];
 
@@ -10,7 +10,7 @@ const render = (args: Args) => html`
 `;
 
 const meta: Meta = {
-  title: 'Styling/focusring',
+  title: 'Styling/Focus Ring',
   tags: ['autodocs'],
   component: 'mdc-subcomponent-focusring',
   render,
