+++ packages/components/src/tsconfig.json
@@ -47,6 +47,5 @@
     "**/*.e2e-test*",
     "**/*.stories*",
     "node_modules",
-    "stories"
   ],
 }
